Update resume.tex
Using a more robust approach with paracol to build the two-column page instead of manually doing it with minipages.
Improved typography with hyphenation and text justification in the larger column, making the paragraphs look more solud and complete. Sloppy ragged text for the smaller column allows for hyphenated words but restrict the line-filling as much.
The item indentation for the lists is too small resulting in zero space between the two columns when a word has a bad width. Changed it so that it looks symmetric and works visually.

Are you sure that the title should be set in the page header? I don't think it is a good idea unless you really want it to appear on the next pages as well (I don't think that would be good either).

If you have to stick with XeTeX, remove the line with microtype.